requirements.txt
streamlit
pandas
plotly
numpy
requests
python-dotenv
azure-data-tables
